v2/requirements.txt
asgiref==3.8.1
Django==4.1.13
django-rest-framework==0.1.0
djangorestframework==3.15.1
djongo==1.3.6
dnspython==2.6.1
drf-yasg==1.21.7
inflection==0.5.1
packaging==24.1
pyaes==1.6.1
pyasn1==0.6.1
pycryptodomex==3.21.0
pymongo==3.12.3
pytz==2024.2
PyYAML==6.0.2
pyzipper==0.3.6
rsa==4.9
setuptools==75.1.0
sqlparse==0.2.4
Telethon==1.37.0
uritemplate==4.1.1
